new components to modularize timeline page
