+++ src/pages/timelinePage/TimelinePage.tsx
@@ -1,13 +1,15 @@
-import { Canvas, useFrame, useThree } from "@react-three/fiber";
-import { Environment, OrbitControls, useProgress } from "@react-three/drei";
-import { useEffect, useMemo, useRef, useState } from "react";
-import * as THREE from "three";
-import { TextureLoader } from "three";
+import { Canvas } from "@react-three/fiber";
+import { Environment, OrbitControls } from "@react-three/drei";
+import { useEffect, useMemo, useState } from "react";
 import { eras } from "../../data/eras";
-import TimelineUI from "../../components/timeline/TimelineUI";
 import Navbar from "../../components/navbar/Navbar";
 import logo from "/images/logoblack.png";
 import { Button } from "../../components/ui/button/Button";
+import IdleCameraOrbit from "../../components/timeline/IdleCameraOrbit";
+import BackgroundCrossfade from "../../components/timeline/BackgroundCrossfade";
+import LoaderOverlay from "../../components/timeline/LoaderOverlay";
+import WithProgressUI from "../../components/timeline/WithProgressUI";
+import { TextureLoader } from "three";
 
 export default function MainScene() {
   // Usar siempre un id existente como inicial para evitar undefined
@@ -57,9 +59,7 @@ export default function MainScene() {
     };
   }, []);
 
-  const isHDR =
-    !!background &&
-    (background.endsWith(".hdr") || background.endsWith(".exr"));
+  const isHDR = !!background && (background.endsWith(".hdr") || background.endsWith(".exr"));
   // El componente de crossfade se encargará de cargar texturas cuando NO sean HDR.
 
   return (
@@ -119,208 +119,4 @@ export default function MainScene() {
       </Button>
     </div>
   );
-}
-
-function WithProgressUI({
-  currentEra,
-  setCurrentEra,
-}: {
-  currentEra: string;
-  setCurrentEra: (id: string) => void;
-}) {
-  const { active } = useProgress();
-  return (
-    <TimelineUI
-      currentEra={currentEra}
-      setCurrentEra={setCurrentEra}
-      loading={active}
-    />
-  );
-}
-
-function LoaderOverlay() {
-  const { active, progress } = useProgress();
-  if (!active) return null;
-  return (
-    <div className="pointer-events-none absolute inset-0 flex items-center justify-center bg-black/60 backdrop-blur-sm z-20">
-      <div className="flex flex-col items-center gap-3 text-white">
-        <div className="w-12 h-12 border-4 border-white/30 border-t-white rounded-full animate-spin" />
-        <span className="text-sm tracking-wide">
-          Cargando {progress.toFixed(0)}%
-        </span>
-      </div>
-    </div>
-  );
-}
-
-function BackgroundCrossfade({ path }: { path: string }) {
-  const currentTexRef = useRef<THREE.Texture | null>(null);
-  const prevTexRef = useRef<THREE.Texture | null>(null);
-  const progressRef = useRef(1);
-  const animatingRef = useRef(false);
-  const lastPathRef = useRef<string | null>(null);
-  const materialRef = useRef<THREE.ShaderMaterial | null>(null);
-  const [revision, setRevision] = useState(0);
-
-  const loadTexture = (url: string) =>
-    new Promise<THREE.Texture>((resolve, reject) => {
-      const loader = new TextureLoader();
-      loader.load(
-        url + "?v=" + Date.now(),
-        (tex) => {
-          tex.colorSpace = THREE.SRGBColorSpace;
-          tex.mapping = THREE.UVMapping;
-          tex.needsUpdate = true;
-          resolve(tex);
-        },
-        undefined,
-        reject
-      );
-    });
-
-  useEffect(() => {
-    if (!path) return;
-    if (lastPathRef.current === path && currentTexRef.current) return;
-    let mounted = true;
-    const previous = currentTexRef.current;
-    if (previous) prevTexRef.current = previous;
-    progressRef.current = previous ? 0 : 1;
-    animatingRef.current = !!previous;
-    loadTexture(path).then((tex) => {
-      if (!mounted) return;
-      currentTexRef.current = tex;
-      // Si no había textura previa, necesitamos re-render para montar el shader
-      if (!previous) {
-        setRevision((r) => r + 1);
-        return;
-      }
-      animatingRef.current = true;
-    });
-    lastPathRef.current = path;
-    return () => {
-      mounted = false;
-    };
-  }, [path]);
-
-  useFrame((_, delta) => {
-    if (!materialRef.current) return;
-    const mat = materialRef.current;
-    if (animatingRef.current) {
-      const speed = 1 / 1.6;
-      progressRef.current = Math.min(1, progressRef.current + delta * speed);
-      if (progressRef.current >= 1) animatingRef.current = false;
-    }
-    const k = progressRef.current;
-    const ease = k < 0.5 ? 4 * k * k * k : 1 - Math.pow(-2 * k + 2, 3) / 2;
-    const blur = Math.sin(ease * Math.PI) * 0.006;
-    mat.uniforms.uTexCurrent.value = currentTexRef.current;
-    mat.uniforms.uTexPrev.value = prevTexRef.current;
-    mat.uniforms.uProgress.value = ease;
-    mat.uniforms.uHasPrev.value = prevTexRef.current ? 1 : 0;
-    mat.uniforms.uBlur.value = blur;
-  });
-
-  const shader = useMemo(
-    () => ({
-      uniforms: {
-        uTexCurrent: { value: null },
-        uTexPrev: { value: null },
-        uProgress: { value: 1 },
-        uHasPrev: { value: 0 },
-        uBlur: { value: 0 },
-      },
-      vertexShader: /* glsl */ `varying vec2 vUv; void main(){ vUv=uv; gl_Position=projectionMatrix*modelViewMatrix*vec4(position,1.0); }`,
-      fragmentShader: /* glsl */ `
-      varying vec2 vUv; 
-      uniform sampler2D uTexCurrent; 
-      uniform sampler2D uTexPrev; 
-      uniform float uProgress; 
-      uniform float uHasPrev; 
-      uniform float uBlur; 
-
-      const float uExposure = 1.9;
-      const float uGamma = 1.3; 
-
-      vec4 sampleBlur(sampler2D tx, vec2 uv, float b){
-        if(b<=0.0001) return texture2D(tx,uv);
-        vec2 o=vec2(b);
-        vec4 c=texture2D(tx,uv)*0.4;
-        c+=texture2D(tx,uv+vec2(o.x,0.0))*0.15;
-        c+=texture2D(tx,uv-vec2(o.x,0.0))*0.15;
-        c+=texture2D(tx,uv+vec2(0.0,o.x))*0.15;
-        c+=texture2D(tx,uv-vec2(0.0,o.x))*0.15;
-        return c; 
-      }
-
-      vec3 srgbToLinear(vec3 c){
-        return pow(max(c,0.0), vec3(uGamma));
-      }
-      vec3 linearToSRGB(vec3 c){
-        return pow(max(c,0.0), vec3(1.0 / uGamma));
-      }
-
-      void main(){
-        float mixF = uProgress;
-        float bP = (1.0 - mixF) * uBlur;
-        float bC = mixF * uBlur;
-        vec4 prev = uHasPrev > 0.5 ? sampleBlur(uTexPrev, vUv, bP) : vec4(0.0);
-        vec4 curr = sampleBlur(uTexCurrent, vUv, bC);
-        vec4 raw = mix(prev, curr, mixF);
-        // Convertimos de sRGB (aprox) a lineal, aplicamos exposición y volvemos a sRGB
-        vec3 linear = srgbToLinear(raw.rgb);
-        linear *= uExposure;
-        vec3 srgb = linearToSRGB(linear);
-        gl_FragColor = vec4(srgb, 1.0);
-      }
-    `,
-    }),
-    []
-  );
-
-  if (!currentTexRef.current && !prevTexRef.current) {
-    return (
-      <mesh scale={[-1, 1, 1]}>
-        <sphereGeometry args={[800, 64, 64]} />
-        <meshBasicMaterial side={THREE.BackSide} color="#000" />
-      </mesh>
-    );
-  }
-
-  return (
-    <mesh key={revision} scale={[-1, 1, 1]}>
-      <sphereGeometry args={[800, 64, 64]} />
-      <shaderMaterial
-        // ref tipado sin usar any
-        ref={(m: THREE.ShaderMaterial | null) => {
-          materialRef.current = m;
-        }}
-        attach="material"
-        args={[shader]}
-        side={THREE.BackSide}
-      />
-    </mesh>
-  );
-}
-
-// Órbita de cámara: rota 360º continuo alrededor del centro (eje Y) muy lentamente.
-function IdleCameraOrbit({
-  active,
-  radius = 3,
-  period = 120,
-}: {
-  active: boolean;
-  radius?: number;
-  period?: number;
-}) {
-  const { camera, clock } = useThree();
-  // Pequeña altura opcional dinámica para hacerlo menos estático
-  useFrame(() => {
-    if (!active) return;
-    const t = clock.getElapsedTime();
-    const angle = (t / period) * Math.PI * 2.0; // ciclo completo cada "period" segundos
-    const y = Math.sin(t * 0.15) * 0.12; // leve vaivén vertical
-    camera.position.set(Math.cos(angle) * radius, y, Math.sin(angle) * radius);
-    camera.lookAt(0, 0, 0);
-  });
-  return null;
 }
